Age and DOB processing

The function first calculates the age based on the present age variable and if unavailable date of birth and enrollment/consent date, end then runs the age_groups function to divide into different categories.

Usage

age_dob_process(
  data.var,
  age.var,
  dob.var,
  date.var,
  lim1,
  lim2,
  lim3 = NULL,
  lim4 = NULL,
  lim5 = NULL
)

Arguments

data.var

Data table

age.var

Variable containing age in years.

dob.var

Variable containing date of birth

date.var

Variable containing enrolment/consent date

lim1

Age limit for categorization

lim2

Age limit for categorization

lim3

Age limit for categorization

lim4

Age limit for categorization

lim5

Age limit for categorization

Value

Returns the data table with a calculated age column and categorized age column added.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(lubridate)
#> 
#> Attaching package: ‘lubridate’
#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union

my_data <- data.frame(
  ID = c(1, 2, 3, 4, 5),
  age = c(25, NA, 30, NA, NA),
  dob = c("1980-01-15", "1995-03-22", "1975-08-10", "1988-11-05", "2000-09-30"),
  date_enrollment = c("2023-05-10", "2022-12-01", "2022-02-18", "2023-02-05", "2021-07-20")
)

# Display the data
print(my_data)
#>   ID age        dob date_enrollment
#> 1  1  25 1980-01-15      2023-05-10
#> 2  2  NA 1995-03-22      2022-12-01
#> 3  3  30 1975-08-10      2022-02-18
#> 4  4  NA 1988-11-05      2023-02-05
#> 5  5  NA 2000-09-30      2021-07-20

# Apply the age_dob_process function to the data
age_dob_process(data.var=my_data, age.var = "age", dob.var = "dob", date.var = "date_enrollment", lim1 = 30, lim2 = 40)
#>   ID age        dob date_enrollment AGE_CALC AGE_CATEG
#> 1  1  25 1980-01-15      2023-05-10 25.00000       <30
#> 2  2  NA 1995-03-22      2022-12-01 27.69610      <NA>
#> 3  3  30 1975-08-10      2022-02-18 30.00000     30-39
#> 4  4  NA 1988-11-05      2023-02-05 34.25051      <NA>
#> 5  5  NA 2000-09-30      2021-07-20 20.80219      <NA>