categorize_result

Categorize a subject as "Positive" or "Negative", based on a variable of interest and a specified threshold. The values are bigger or equal to the threshold are categorized as "Positive", others "Negative". A list of thresholds can be provided.

Usage

categorize_result(dataframe, variable, thresholds)

Arguments

dataframe

The data table containing test results

variable

A continuous variable

thresholds

The thresholds to be applied

Value

Data table with the categorization added

Examples

df <- data.frame(results = c(sample(1:100, 50, replace=TRUE)))
categorize_result(dataframe = df, variable = "results", thresholds = 50)
#>    results Result50
#> 1       32 Negative
#> 2       53 Positive
#> 3       70 Positive
#> 4        9 Negative
#> 5       22 Negative
#> 6       10 Negative
#> 7       32 Negative
#> 8       34 Negative
#> 9       17 Negative
#> 10      52 Positive
#> 11      56 Positive
#> 12      22 Negative
#> 13      95 Positive
#> 14       1 Negative
#> 15      70 Positive
#> 16      34 Negative
#> 17      51 Positive
#> 18      46 Negative
#> 19      65 Positive
#> 20      25 Negative
#> 21      37 Negative
#> 22       8 Negative
#> 23      33 Negative
#> 24      96 Positive
#> 25      83 Positive
#> 26      92 Positive
#> 27      28 Negative
#> 28      48 Negative
#> 29      75 Positive
#> 30      90 Positive
#> 31      35 Negative
#> 32       2 Negative
#> 33      87 Positive
#> 34       3 Negative
#> 35      38 Negative
#> 36      68 Positive
#> 37      65 Positive
#> 38       6 Negative
#> 39      22 Negative
#> 40      19 Negative
#> 41       2 Negative
#> 42      64 Positive
#> 43      40 Negative
#> 44      65 Positive
#> 45      28 Negative
#> 46      97 Positive
#> 47      71 Positive
#> 48      12 Negative
#> 49      54 Positive
#> 50      41 Negative