compare_sen_spe

Because these are two independent groups, a Z-score is calculated from the two proportions (i.e. sensitivities or specificities) to be compared. By default, the test is two tailed, but lower or upper tail can be specified in parameters.

Usage

compare_sen_spe(s1, s2, n1, n2, two.sided = TRUE)

Arguments

s1

Sensitivity/specificity of the first group

s2

Sensitivity/specificity of the second group

n1

Total number of confirmed positives/negatives respectively in the first group

n2

Total number of confirmed positives/negatives respectively in the second group

two.sided

If TRUE, equivalence is tested, if FALSE, directional inferences are made

Value

A dataframe containing Z-Score and the P-value

References

Zhou XH, Obuchowski NA and McClish DK. Statistical Methods in Diagnostic Medicine. 2011;2:193-228

Examples

result.df <- data.frame(Tests = c("Test1", "Test2"), TP = c(82, 140), FN = c(18, 60))
result.df$N <- result.df$TP +result.df$FN
result.df$sen <- result.df$TP/result.df$N
compare_sen_spe(s1 = result.df$sen[1], s2 = result.df$sen[2], n1 = result.df$N[1], n2 = result.df$N[2],two.sided = TRUE)
#>   Statistic       Value
#> 1         Z -2.38761651
#> 2         P  0.01695803