multi_sen_spe for Table Display

Calculates and displays sensitivity and specificity with confidence intervals from a data frame for multiple index tests and the reference test. To be used for displaying the result in a table.

Usage

multi_sen_spe(data_var, list_index, ref, conf.level = 0.95, index_names = NULL)

Arguments

data_var

Data table containing tests results

list_index

A list of index tests with "Negative" and "Positive" results

ref

The reference test with "Negative" and "Positive" results

conf.level

The confidence level, 1-alpha – default 95%

index_names

An optional list containing names of the index tests

Value

A data table with TP/FP/TN/FN, sensitivity, specificity and confidence intervals for all index tests specified in the list

Examples

library(finddataanalysis)
df <- data.frame(Index1 = c(rep("Positive", 20), rep("Negative", 20)), Index2 = c(rep("Positive", 5), rep("Negative", 35)), Index3 = c(rep("Positive", 9), rep("Negative", 31)), Reference = c(rep("Positive", 10), rep("Negative", 30)))

multi_sen_spe(data_var = df, list_index = c("Index1", "Index2", "Index3"), ref = "Reference", conf.level = 0.95, index_names = c("a", "b", "c"))
#>    N TP FP FN TN Sensitivity Sensitivity [95%CI] Specificity
#> a 40 10 10  0 20         100       [72.25-100.0]       66.67
#> b 40  5  0  5 30          50       [23.66-76.34]         100
#> c 40  9  0  1 30          90       [59.58-98.21]         100
#>   Specificity [95%CI] Balanced Accuracy B. Accuracy [95%CI] DOR DOR [95%CI] PPV
#> a       [48.78-80.77]             83.34       [60.52-90.38] Inf   [NaN-Inf]  50
#> b       [88.65-100.0]                75       [56.15-88.17] Inf   [NaN-Inf] 100
#> c       [88.65-100.0]                95        [74.11-99.1] Inf   [NaN-Inf] 100
#>     PPV [95%CI]   NPV   NPV [95%CI] Accuracy Accuracy [95%CI]
#> a [29.93-70.07]   100 [83.89-100.0]       75    [59.81-85.81]
#> b [56.55-100.0] 85.71 [70.62-93.74]     87.5    [73.89-94.54]
#> c [70.09-100.0] 96.77 [83.81-99.43]     97.5    [87.12-99.56]