multi_sen_spe with DT table output

Executes multi_sen_spe function and creates a nicely formatted table output with download options using DT package

Usage

multi_sen_spe_dt_out(
  data_var,
  list_index,
  ref,
  conf.level = 0.95,
  index_names = NULL,
  file_name = "performance_eval"
)

Arguments

data_var

Data table containing tests results

list_index

A list of index tests with "Negative" and "Positive" results

ref

The reference test with "Negative" and "Positive" results

conf.level

The confidence level, 1-alpha – default 95%

index_names

An optional list containing names of the index tests

file_name

An optional text containing the name of the file that will be downloaded

Value

A formatted data table displaying performance characteristics

Examples

df <- data.frame(Index1 = c(rep("Positive", 20), rep("Negative", 20)), Index2 = c(rep("Positive", 5), rep("Negative", 35)), Index3 = c(rep("Positive", 9), rep("Negative", 31)), Reference = c(rep("Positive", 10), rep("Negative", 30)))

multi_sen_spe_dt_out(data_var = df, list_index = c("Index1", "Index2", "Index3"), ref = "Reference", conf.level = 0.95, file_name = "ZX001_performance", index_names = c("a", "b", "c"))