multi_sen_spe_forest for forest plots

Calculates and displays sensitivity and specificity with confidence intervals from a data frame for multiple index tests and the reference test. To be used for displaying the result in a table.

Usage

multi_sen_spe_forest(
  data_var,
  list_index,
  ref,
  conf.level = 0.95,
  index_names = NULL
)

Arguments

data_var

Data table containing test results

list_index

A list of index test with "Negative" and "Positive" results

ref

The reference test with "Negative" and "Positive" results

conf.level

The confidence level, 1-alpha – default 95%

index_names

An optional list containing names of the index tests

Value

A data table with TP/FP/TN/FN, sensitivity, specificity and confidence intervals, to be fed into forestplot function.

Examples

df <- data.frame(Index1 = c(rep("Positive", 20), rep("Negative", 20)), Index2 = c(rep("Positive", 5), rep("Negative", 35)), Index3 = c(rep("Positive", 9), rep("Negative", 31)), Reference = c(rep("Positive", 10), rep("Negative", 30)))

multi_sen_spe_forest(data_var = df, list_index = c("Index1", "Index2", "Index3"), ref = "Reference", conf.level = 0.95, index_names = c("ate","FNDH", "DXCb"))
#>    N TP FP FN TN Sensitivity Specificity SensLower SensUpper SpeLower SpeUpper
#> 1 40 10 10  0 20         100       66.67     72.25    100.00    48.78    80.77
#> 2 40  5  0  5 30          50      100.00     23.66     76.34    88.65   100.00
#> 3 40  9  0  1 30          90      100.00     59.58     98.21    88.65   100.00
#>   Balanced_Accuracy BAlower BAupper DOR DORUpper DORLower PPV PPVLower PPVUpper
#> 1             83.34   60.52   90.38 Inf      Inf      NaN  50    29.93    70.07
#> 2             75.00   56.15   88.17 Inf      Inf      NaN 100    56.55   100.00
#> 3             95.00   74.11   99.10 Inf      Inf      NaN 100    70.09   100.00
#>      NPV NPVLower NPVUpper Accuracy ACCLower ACCUpper Test
#> 1 100.00    83.89   100.00     75.0    59.81    85.81  ate
#> 2  85.71    70.62    93.74     87.5    73.89    94.54 FNDH
#> 3  96.77    83.81    99.43     97.5    87.12    99.56 DXCb