performance_eval_auto

Calculates diagnostic accuracy characteristics and generates report quality tables and figures. Executes multi_sen_spe_out_forest, multi_sen_spe_dt_out, and returns a list object with two forest plots and a DT data table for display.The function works with wide and long formats.

Usage

performance_eval_auto(
  data_var,
  list_index,
  ref,
  conf.level = 0.95,
  index_names = NULL,
  labels = "Test",
  forest_plot = TRUE,
  table_output = TRUE,
  file_name = "performance_eval",
  data_long = FALSE,
  group_var = NULL
)

Arguments

data_var

Data table containing tests results

list_index

A list of index tests with "Negative" and "Positive" results. If the data is in long format, the name of the column where the results are stored.

ref

The reference test with "Negative" and "Positive" results

conf.level

The confidence level, 1-alpha – default 95%

index_names

An optional list containing names of the index tests. Not used with long data format.

labels

An optional text for the axis labels in forest plots.

forest_plot

Should the forest plots be generated? Default TRUE

table_output

Should a table output be generated? Default TRUE

file_name

An optional text for the file name that will be available for download when the table output is generated

data_long

Is the data in the long format? Default FALSE

group_var

The variable name where the groups are defined. Enter without quotation marks (e.g. ColName instead of "ColName"). Not used with wide data format.

Value

performance evaluation results as plots, formatted table, or simple data frame

Examples

df <- data.frame(Index1 = c(rep("Positive", 20), rep("Negative", 20)), Index2 = c(rep("Positive", 5), rep("Negative", 35)), Index3 = c(rep("Positive", 9), rep("Negative", 31)), Reference = c(rep("Positive", 10), rep("Negative", 30)))

# All outputs
eval_output <- performance_eval_auto(data_var = df, list_index = c("Index1", "Index2", "Index3"), ref = "Reference", conf.level = 0.95, index_names = c("Tgs1", "AFD", "SimpleDx"), labels = "Test", forest_plot = TRUE, table_output = TRUE, file_name = "MyEvaluationExample")
eval_output$sen_plot 

eval_output$spe_plot

eval_output$table

# Forest plot outputs
eval_output_only_forest <- performance_eval_auto(data_var = df, list_index = c("Index1", "Index2", "Index3"), ref = "Reference", conf.level = 0.95, index_names = c("Tgs1", "AFD", "SimpleDx"), labels = "Test", forest_plot = TRUE, table_output = FALSE)
eval_output_only_forest$sen_plot

eval_output_only_forest$spe_plot

eval_output_only_forest$table #NULL
#> NULL

# A simple data frame output where the table is not formatted. This form is computer friendly.
eval_output_simple_df <- performance_eval_auto(data_var = df, list_index = c("Index1", "Index2", "Index3"), ref = "Reference", conf.level = 0.95, index_names = c("Tgs1", "AFD", "SimpleDx"), labels = "Test", forest_plot = FALSE, table_output = FALSE)

eval_output_simple_df
#>    N TP FP FN TN Sensitivity Specificity SensLower SensUpper SpeLower SpeUpper
#> 1 40 10 10  0 20         100       66.67     72.25    100.00    48.78    80.77
#> 2 40  5  0  5 30          50      100.00     23.66     76.34    88.65   100.00
#> 3 40  9  0  1 30          90      100.00     59.58     98.21    88.65   100.00
#>   Balanced_Accuracy BAlower BAupper DOR DORUpper DORLower PPV PPVLower PPVUpper
#> 1             83.34   60.52   90.38 Inf      Inf      NaN  50    29.93    70.07
#> 2             75.00   56.15   88.17 Inf      Inf      NaN 100    56.55   100.00
#> 3             95.00   74.11   99.10 Inf      Inf      NaN 100    70.09   100.00
#>      NPV NPVLower NPVUpper Accuracy ACCLower ACCUpper     Test
#> 1 100.00    83.89   100.00     75.0    59.81    85.81     Tgs1
#> 2  85.71    70.62    93.74     87.5    73.89    94.54      AFD
#> 3  96.77    83.81    99.43     97.5    87.12    99.56 SimpleDx

# Performance Evaluation by Groups
data(my_dataset)
head(my_dataset)
#>         ID Test_Name   Result  RefTest
#> 1 01STUDID       ABC Negative Negative
#> 2 01STUDID       DST Negative Negative
#> 3 01STUDID       HCG Negative Negative
#> 4 01STUDID      LLJK Negative Negative
#> 5 01STUDID      ddfs Negative Negative
#> 6 01STUDID Weird Lab Negative Negative
eval_output <- performance_eval_auto(data_var = my_dataset, list_index = "Result", ref = "RefTest", conf.level = 0.95, labels = "Test", forest_plot = FALSE, table_output = TRUE, file_name = "MyEvaluationExample", data_long = TRUE, group_var = Test_Name )