Regroup axillary temperature values

Creates a new categorical variable based on the body temperature.

Usage

temperature_groups(data.var, temperature)

Arguments

data.var

Data table

temperature

Variable containing body temperature in Celsius

Value

Creates a new column frame with body temperature ranges

Examples

df = data.frame(ID = c(1:20), Temperature = runif(20, 36, 42))
temperature_groups(data.var = df, temperature = "Temperature")
#>    ID Temperature TEMP_CATEG
#> 1   1    41.50596        ≥39
#> 2   2    41.83707        ≥39
#> 3   3    40.91449        ≥39
#> 4   4    41.41754        ≥39
#> 5   5    39.48820        ≥39
#> 6   6    40.63805        ≥39
#> 7   7    41.97074        ≥39
#> 8   8    40.26583        ≥39
#> 9   9    37.28966    37-37.9
#> 10 10    37.75055    37-37.9
#> 11 11    40.33056        ≥39
#> 12 12    41.19969        ≥39
#> 13 13    37.43072    37-37.9
#> 14 14    36.02698    36-36.9
#> 15 15    41.66110        ≥39
#> 16 16    38.62882    38-38.9
#> 17 17    40.50362        ≥39
#> 18 18    40.00689        ≥39
#> 19 19    38.44784    38-38.9
#> 20 20    38.10749    38-38.9